Implement C99 T6 conformance corrections

+++ include/stdlib.h
@@ -12,7 +12,7 @@
 #define EXIT_SUCCESS 0
 #define EXIT_FAILURE 1
 #define RAND_MAX 32767
-#define MB_CUR_MAX 1
+#define MB_CUR_MAX ((size_t)1)
 
 /** @brief Quotient and remainder from integer division. */
 typedef struct { int quot; int rem; } div_t;
